New win logic
The player with the shortest path for now wins.
New players are forbidden to join a game which has already a winner.
Everybody who can reach the goal in a shorter path is the new winner.

There may be points for winning, even when one is not the winner
anymore.