Fix #4
The race condition between the temlate rendering of game.html and
the connection of the websockets is solved by applying the visit
count on every join, regardless of whether the player exists or not.

The template renders the player list including the joining player,
which results in the joinHandler to abort and do nothing. Now
the join handler sets the visits correctly.